feat: separate dockers
