--- Dockerfile
@@ -1,110 +1,0 @@
-ARG image
-
-FROM alpine:latest AS base
-RUN apk add --update --no-cache curl docker git openssh-client
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM golang:1.15-alpine AS golang-1-15
-RUN apk add --update --no-cache curl docker git openssh-client
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM golang:1.16-alpine AS golang-1-16
-RUN apk add --update --no-cache curl docker git openssh-client
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM node:15-alpine AS node-15
-RUN apk add --update --no-cache curl docker python3 alpine-sdk
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM node:14-alpine AS node-14
-RUN apk add --update --no-cache curl docker python3 alpine-sdk
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM node:13-alpine AS node-13
-RUN apk add --update --no-cache curl docker python3 alpine-sdk
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM python:3.9-alpine AS python
-RUN apk add --update --no-cache curl docker
-RUN mkdir -p ~/.docker/cli-plugins && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM maven:3.8.1-openjdk-16-slim AS java-16
-RUN mkdir -p ~/.docker/cli-plugins && \
-	curl https://download.docker.com/linux/static/stable/x86_64/docker-20.10.7.tgz -o docker-20.10.7.tgz && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan && \
-    tar xzvf docker-20.10.7.tgz && \
-    cp docker/* /usr/bin/
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM maven:3.8.1-openjdk-15-slim AS java-15
-RUN mkdir -p ~/.docker/cli-plugins && \
-	curl https://download.docker.com/linux/static/stable/x86_64/docker-20.10.7.tgz -o docker-20.10.7.tgz && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan && \
-    tar xzvf docker-20.10.7.tgz && \
-    cp docker/* /usr/bin/
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM maven:3.8.1-openjdk-11-slim AS java-11
-RUN mkdir -p ~/.docker/cli-plugins && \
-	curl https://download.docker.com/linux/static/stable/x86_64/docker-20.10.7.tgz -o docker-20.10.7.tgz && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan && \
-    tar xzvf docker-20.10.7.tgz && \
-    cp docker/* /usr/bin/
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM maven:3.8.1-openjdk-8-slim AS java-8
-RUN mkdir -p ~/.docker/cli-plugins && \
-	curl https://download.docker.com/linux/static/stable/x86_64/docker-20.10.7.tgz -o docker-20.10.7.tgz && \
-    curl https://github.com/docker/scan-cli-plugin/releases/latest/download/docker-scan_linux_amd64 -L -s -S -o ~/.docker/cli-plugins/docker-scan && \
-    chmod +x ~/.docker/cli-plugins/docker-scan && \
-    tar xzvf docker-20.10.7.tgz && \
-    cp docker/* /usr/bin/
-COPY bin/linux-amd64/winch /usr/local/bin/winch
-COPY entrypoint.sh /entrypoint.sh
-ENTRYPOINT ["/entrypoint.sh"]
-
-FROM ${image}
