+++ cmd/winch/commands/ci.go
@@ -20,8 +20,8 @@ import (
 	"context"
 	"fmt"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch"
-	"github.com/winchci/winch/config"
+	winch "github.com/winchci/winch/pkg"
+	"github.com/winchci/winch/pkg/config"
 	"os"
 )
