+++ cmd/winch/commands/docker_publish.go
@@ -20,9 +20,9 @@ import (
 	"context"
 	"fmt"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch"
-	"github.com/winchci/winch/config"
-	"github.com/winchci/winch/docker"
+	winch "github.com/winchci/winch/pkg"
+	"github.com/winchci/winch/pkg/config"
+	"github.com/winchci/winch/pkg/docker"
 )
 
 func dockerPublish(ctx context.Context) error {
