+++ cmd/winch/commands/generate_assets.go
@@ -21,8 +21,8 @@ import (
 	"github.com/shurcooL/httpfs/filter"
 	"github.com/shurcooL/vfsgen"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch"
-	"github.com/winchci/winch/config"
+	winch "github.com/winchci/winch/pkg"
+	"github.com/winchci/winch/pkg/config"
 	"net/http"
 	"os"
 	"path/filepath"
