+++ cmd/winch/commands/generate_dockerfile.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"github.com/golang/protobuf/proto"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch/config"
+	"github.com/winchci/winch/pkg/config"
 	"github.com/winchci/winch/templates"
 	"os"
 	"path/filepath"
