+++ cmd/winch/commands/generate_environment.go
@@ -21,7 +21,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch/config"
+	"github.com/winchci/winch/pkg/config"
 	"io/ioutil"
 )
