+++ cmd/winch/commands/generate_version.go
@@ -20,8 +20,8 @@ import (
 	"context"
 	"encoding/json"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch"
-	"github.com/winchci/winch/config"
+	winch "github.com/winchci/winch/pkg"
+	"github.com/winchci/winch/pkg/config"
 	"github.com/winchci/winch/templates"
 	"gopkg.in/yaml.v3"
 	"io/ioutil"
