+++ cmd/winch/commands/init.go
@@ -21,8 +21,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch"
-	"github.com/winchci/winch/config"
+	winch "github.com/winchci/winch/pkg"
+	"github.com/winchci/winch/pkg/config"
 	"gopkg.in/yaml.v3"
 	"io/ioutil"
 	"os"
