+++ cmd/winch/commands/root.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch/config"
+	"github.com/winchci/winch/pkg/config"
 	"github.com/winchci/winch/version"
 	"log"
 	"os"
