+++ cmd/winch/commands/transom_publish.go
@@ -21,9 +21,9 @@ import (
 	"fmt"
 	"github.com/mholt/archiver/v3"
 	"github.com/spf13/cobra"
-	"github.com/winchci/winch"
-	"github.com/winchci/winch/config"
-	"github.com/winchci/winch/transom"
+	winch "github.com/winchci/winch/pkg"
+	"github.com/winchci/winch/pkg/config"
+	"github.com/winchci/winch/pkg/transom"
 	"github.com/winchci/winch/version"
 	"io/ioutil"
 	"os"
