+++ pkg/namer.go
@@ -14,7 +14,7 @@ You should have received a copy of the GNU General Public License along with thi
 see <https://www.gnu.org/licenses/>.
 */
 
-package winch
+package pkg
 
 import (
 	"context"
